+++ Changelog.md
@@ -5,6 +5,7 @@
 - Add new routes for `update`, `show`, and `index` actions of the Sections API Controller (#6955)
 - Enable the deletion of Grade Entry Forms that have no grades (#6915)
 - Fixed login_spec.rb flaky test on GitHub Actions run (#6966)
+- Enable bulk download of print PDFs for an assignments (#6998)
 
 ## [v2.4.6]
 - Disallow students from uploading .git file and .git folder in their repository (#6963)